requirements.txt
airflow-provider-duckdb==0.0.2
astro-sdk-python==1.5.0
# duckdb==0.7.1
duckdb==1.0.0
minio==7.1.13
pandas==1.5.3
streamlit==1.23.0
aiohttp